Started op_loadout framework
Implemented framework for the LOADOUT system on the backend. The 'loadout key' is in place, and designates which weapons will go to which slots (although due to lack of weapons this is far from complete.)
Next is to figure out a way for players to interact and establish their chosen LOADOUT. Attempted to do this through a custom cvar method, although this ran into a dead end because of a lack capabilities of Ironwail to do cvars in this way. That code has been reverted here, although the baseline LOADOUT system is still in place.