unstable-majoroverhaul
Most of the work of moving code around is complete. The code compiles only with the 'doors warning' that existed in the QRRQC code.
It appears though, we've stumbled upon our 'phantom' yet again, though with a slightly different symptom. Now it appears to be causing a buffer overflow into the logic of func_train operations.
If not, then that is the result of the structure of the debugging code that was used to identify the error.