Major Update - crashing solved
Apparently the phantom was a result of a large amount of mixed defs files. To solve this, the defs.qc was restored and used as the only defs file. Changes to related qc files were updated to no longer reflect the split file setup.
A new, well researched split defs setup is planned, making use of #define.

This version appears stable, and functioning, outside of known unaddressed monster behaviors that were earlier removed in preparation for modularity.